This commit essentially "closes" 26.
It's based on the following a user cast 1 vote with a value of
-1, 0, and 1.

It's actually nit exactly like that. There is no vote with an explicit value of 0.

A user can cast one vote only against any give proposal, with a value of 'up' or 'down' which, from the proposal's point of view, should it have one, equals to a 1 or -1.

If, however, a voter has already cast their vote against the proposal, then this new click simply zeroes all the values in the user's vote record. A consequent vote will count again as the expected 1 or -1. I am assuming that this IS the expected behaviour. I am, leaving this issue open nevertheless with a 'testing awaiting' label attached to it to give us time to confirm that this is what we want, and, after testing, that this what it does.