Nice "toast" messages give users feedback when voting, e.g. "You've already voted this down" or "You need to be logged in to vote" etc.
After hours of looking around ad at very many promising notifiers, and each one of t hem had something uniquely nice about it but.., I finally found this which seems to support all we want, including dragging it out the window with oer's thumb on a mobile or tablet. I have not tested this, will do and report.
 This relates to #24 which I am not closing since not all responses have been covered yet.