Category filtering is now fully functioning. Unlike the implementation planned originally, this uses native React ways of achieving this. It sets the selected flags on the categories, themselves store in the application's state. From that point on, any change in that state will trigger a 'tree shakeup' and a diff. and consequently a smooth refresh of the main area, the listing of the proposals.
Note that whilst I looked at listening at events external to the Proposal component's tree, it transpired that one can attach portals to a component which can live anywhere in the page. It's a much nicer way of achieving this and, after biting the original bullet, it's actually much easier as it belongs to the same cluster of logic as the main component

In addition, the buttons now don't turn into outlines when unselected, they simply appear faded.

Proudly closed #4.